First commited as before_4900_rev1

+++ org.eclipse.jgit/src/org/eclipse/jgit/transport/PackParser.java
@@ -58,10 +58,10 @@ import java.util.concurrent.TimeUnit;
 import java.util.zip.DataFormatException;
 import java.util.zip.Inflater;
 
+import org.eclipse.jgit.JGitText;
 import org.eclipse.jgit.errors.CorruptObjectException;
 import org.eclipse.jgit.errors.MissingObjectException;
 import org.eclipse.jgit.errors.TooLargeObjectInPackException;
-import org.eclipse.jgit.internal.JGitText;
 import org.eclipse.jgit.lib.AnyObjectId;
 import org.eclipse.jgit.lib.BatchingProgressMonitor;
 import org.eclipse.jgit.lib.Constants;
@@ -506,17 +506,17 @@ public abstract class PackParser {
 				resolveDeltas(resolving);
 				if (entryCount < objectCount) {
 					if (!isAllowThin()) {
-						throw new IOException(MessageFormat.format(
-								JGitText.get().packHasUnresolvedDeltas,
-								Long.valueOf(objectCount - entryCount)));
+						throw new IOException(MessageFormat.format(JGitText
+								.get().packHasUnresolvedDeltas,
+								(objectCount - entryCount)));
 					}
 
 					resolveDeltasWithExternalBases(resolving);
 
 					if (entryCount < objectCount) {
-						throw new IOException(MessageFormat.format(
-								JGitText.get().packHasUnresolvedDeltas,
-								Long.valueOf(objectCount - entryCount)));
+						throw new IOException(MessageFormat.format(JGitText
+								.get().packHasUnresolvedDeltas,
+								(objectCount - entryCount)));
 					}
 				}
 				resolving.endTask();
@@ -573,14 +573,13 @@ public abstract class PackParser {
 			break;
 		default:
 			throw new IOException(MessageFormat.format(
-					JGitText.get().unknownObjectType,
-					Integer.valueOf(info.type)));
+					JGitText.get().unknownObjectType, info.type));
 		}
 
 		if (!checkCRC(oe.getCRC())) {
 			throw new IOException(MessageFormat.format(
-					JGitText.get().corruptionDetectedReReadingAt,
-					Long.valueOf(oe.getOffset())));
+					JGitText.get().corruptionDetectedReReadingAt, oe
+							.getOffset()));
 		}
 
 		resolveDeltas(visit.next(), info.type, info, progress);
@@ -599,8 +598,7 @@ public abstract class PackParser {
 
 			default:
 				throw new IOException(MessageFormat.format(
-						JGitText.get().unknownObjectType,
-						Integer.valueOf(info.type)));
+						JGitText.get().unknownObjectType, info.type));
 			}
 
 			byte[] delta = inflateAndReturn(Source.DATABASE, info.size);
@@ -612,7 +610,7 @@ public abstract class PackParser {
 			if (!checkCRC(visit.delta.crc))
 				throw new IOException(MessageFormat.format(
 						JGitText.get().corruptionDetectedReReadingAt,
-						Long.valueOf(visit.delta.position)));
+						visit.delta.position));
 
 			objectDigest.update(Constants.encodedTypeString(type));
 			objectDigest.update((byte) ' ');
@@ -651,8 +649,7 @@ public abstract class PackParser {
 
 			default:
 				throw new IOException(MessageFormat.format(
-						JGitText.get().unknownObjectType,
-						Integer.valueOf(typeCode)));
+						JGitText.get().unknownObjectType, typeCode));
 			}
 	}
 
@@ -684,7 +681,7 @@ public abstract class PackParser {
 		while ((c & 0x80) != 0) {
 			c = readFrom(Source.DATABASE);
 			hdrBuf[hdrPtr++] = (byte) c;
-			sz += ((long) (c & 0x7f)) << shift;
+			sz += (c & 0x7f) << shift;
 			shift += 7;
 		}
 		info.size = sz;
@@ -716,8 +713,7 @@ public abstract class PackParser {
 
 		default:
 			throw new IOException(MessageFormat.format(
-					JGitText.get().unknownObjectType,
-					Integer.valueOf(info.type)));
+					JGitText.get().unknownObjectType, info.type));
 		}
 		return info;
 	}
@@ -835,7 +831,7 @@ public abstract class PackParser {
 		final long vers = NB.decodeUInt32(buf, p + 4);
 		if (vers != 2 && vers != 3)
 			throw new IOException(MessageFormat.format(
-					JGitText.get().unsupportedPackVersion, Long.valueOf(vers)));
+					JGitText.get().unsupportedPackVersion, vers));
 		objectCount = NB.decodeUInt32(buf, p + 8);
 		use(hdrln);
 
@@ -896,7 +892,7 @@ public abstract class PackParser {
 		while ((c & 0x80) != 0) {
 			c = readFrom(Source.INPUT);
 			hdrBuf[hdrPtr++] = (byte) c;
-			sz += ((long) (c & 0x7f)) << shift;
+			sz += (c & 0x7f) << shift;
 			shift += 7;
 		}
 
@@ -956,9 +952,8 @@ public abstract class PackParser {
 		}
 
 		default:
-			throw new IOException(
-					MessageFormat.format(JGitText.get().unknownObjectType,
-							Integer.valueOf(typeCode)));
+			throw new IOException(MessageFormat.format(
+					JGitText.get().unknownObjectType, typeCode));
 		}
 	}
 
@@ -1043,8 +1038,7 @@ public abstract class PackParser {
 
 			if (info.type != Constants.OBJ_BLOB)
 				throw new IOException(MessageFormat.format(
-						JGitText.get().unknownObjectType,
-						Integer.valueOf(info.type)));
+						JGitText.get().unknownObjectType, info.type));
 
 			ObjectStream cur = readCurs.open(obj, info.type).openStream();
 			try {
